+++ lisp/eab-desktop.el
@@ -9,12 +9,37 @@
 
 (require 'desktop)
 
-(if (or (eab/ondaemon (eab/server-P)) (eab/ondaemon (eab/server-C)) (eab/ondaemon "serverM"))
+(if (or
+     (eab/ondaemon (eab/server-P))
+     (eab/ondaemon (eab/server-C))
+     (eab/ondaemon "serverM"))
     (add-hook 'desktop-save-hook (lambda ()
                                    ;; (eab/wg-update-all-workgroups)
                                    (eab/wg-save eab/workgroups-save)
                                    )))
-;; see also eab-appt.el
+
+;; Activate appointments so we get notifications
+(when (and (eab/ondaemon (eab/server-P)) (not noninteractive))
+  (appt-activate t)
+  (run-at-time nil nil (lambda () (message "%s" "hello")))
+  (run-at-time 60 3600 (lambda ()
+                         (interactive)
+                         (if desktop-file-modtime
+                             (desktop-save (eab/desktop-dir))))))
+
+;; edit timer-idle-list
+(when (eab/ondaemon (eab/server-P))
+  (run-with-idle-timer
+   15 't (lambda () (interactive) (save-some-buffers 't))))
+
+(when (eab/ondaemon (eab/server-P))
+  (run-with-idle-timer
+   5 't (lambda () (interactive) (eab/update-query-on-idle 'eab/org-ql-H-query))))
+
+(when (eab/onhost "chronos-emacs")
+  (run-with-idle-timer
+   100 't (lambda () (interactive) (eab/update-gr-status-on-idle))))
+
 
 (defun eab/load-desktop ()
   ;; TODO don't setup defadvice wg-switch-to-workgroup before it