+++ lisp/eab-ui.el
@@ -4,7 +4,7 @@
 ;;
 ;; Author: artscan@list.ru
 ;; Keywords: 
-;; Requirements: ansi-color, autorevert, async
+;; Requirements: ansi-color, async
 ;; Status: not intended to be distributed yet
 
 (eab/bind-path custom-file)
@@ -14,72 +14,11 @@
 (eab/bind-path save-place-file)
 (eab/bind-path url-configuration-directory)
 (eab/bind-path source-directory)
-(eab/bind-path recentf-save-file)
 (eab/bind-path custom-file)
 
 (setq find-function-C-source-directory source-directory)
-(setq recentf-max-saved-items 200)
-
-(defun eab/recentf-eabpool ()
-  (mapcar
-   (lambda (x)
-     (replace-regexp-in-string
-      (concat "^" (regexp-quote "~/")) ""
-      (replace-regexp-in-string
-       (concat "^" (regexp-quote "~/pnt/lion/")) "" x)))
-   (seq-filter
-    (lambda (x)
-      (file-exists-p x))
-    (seq-filter
-     (lambda (x)
-       (or
-        (let ((s "~/git/"))
-          (and (length> x (length s)) (string= (substring x 0 (length s)) s)))
-        (let ((s "~/pnt/lion/data/"))
-          (and (length> x (length s)) (string= (substring x 0 (length s)) s)))))
-     (seq-filter
-      (lambda (x)
-        (let ((s "~/git/org"))
-          (and (length> x (length s)) (not (string= (substring x 0 (length s)) s))))) recentf-list)))))
-
-(add-hook 'after-save-hook
-          (lambda ()
-            (f-write-text
-             (string-join
-              (eab/recentf-eabpool) "\n")
-             'utf-8
-             (concat recentf-save-file "-eabpool"))))
-
-(require 'autorevert)
-(global-auto-revert-mode)
-;; TODO with notify my workflow breaks
-;; not all org buffers auto reverts, i.e. a bug?
-;; (setq auto-revert-use-notify 't)
-;; (setq auto-revert-avoid-polling 't)
-(setq auto-revert-use-notify nil)
-;; (setq auto-revert-remote-files 't) it breaks magit-status buffer
-
-(require 'async)
-
-(require 'ansi-color)
 
 (setq ring-bell-function 'ignore)
-
-;; see eab-compile.el
-(defun colorize-compilation-buffer ()
-  (read-only-mode 'toggle)
-  (ansi-color-apply-on-region (point-min) (point-max))
-  (read-only-mode 'toggle))
-
-;; TODO why concrete buffer-name only?
-(defadvice display-message-or-buffer (before ansi-color activate)
-  "Process ANSI color codes in shell output."
-  (let ((buf (ad-get-arg 0)))
-    (and (bufferp buf)
-         (string= (buffer-name buf) "*Shell Command Output*")
-         (with-current-buffer buf
-           (ansi-color-apply-on-region (point-min) (point-max))))))
-
 
 (setq disabled-command-function nil)
 (fset 'yes-or-no-p 'y-or-n-p)
@@ -88,10 +27,6 @@
 ;; DONE теперь не работают TeX-master "main" в LaTeX-mode
 ;; уже привык их задавать вручную
 (setq enable-local-variables nil)
-
-(setq ediff-window-setup-function 'ediff-setup-windows-plain)
-(setq ediff-diff-options "-w")
-(setq ediff-split-window-function 'split-window-horizontally)
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 
@@ -159,20 +94,8 @@
         (delete-frame))
     ))
 
-(setq browse-url-browser-function (quote browse-url-firefox))
-(setq browse-url-firefox-program "/usr/local/bin/browser-remote")
 
 (setq eww-search-prefix "https://duckduckgo.com/html/?q=")
-
-(defun google ()
-  "Google the selected region if any, display a query prompt otherwise."
-  (interactive)
-  (browse-url
-   (concat
-    "https://www.google.com/search?q="
-    (url-hexify-string (if mark-active
-                           (buffer-substring (region-beginning) (region-end))
-                         (read-string "Google: "))))))
 
 (setq page-delimiter "^$")
 ;; (setq paragraph-start ...)